collector/pyproject.toml
[project]
name = "collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "kiwipiepy>=0.22.2",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
]
